Add initial README file with project overview and key information
This commit adds the initial README file for the Africa Fast Ferry Ltd GitHub repository. The README provides an overview of the project, including:

Purpose and scope of the Africa Fast Ferry tokenization initiative

Description of AFFEQ and AFFR tokens

Compliance with ERC-1400 / ERC-3643 standards

Integration with KYC/AML and revenue distribution in USDC

Planned use of the contract for secure, compliant fundraising under SEC Reg D Rule 506(c) and Regulation S

This file sets the foundation for project documentation and will be expanded as development progresses.